pyproject.toml
[project]
name = "taskmanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.27.2",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "py-trello>=0.20.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
]
